registro de retiradas

+++ app/Http/Controllers/Root/UserController.php
@@ -5,6 +5,8 @@ namespace App\Http\Controllers\Root;
 use App\Http\Controllers\Controller;
 use App\{Company, User};
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+use Spatie\Permission\Exceptions\UnauthorizedException;
 
 class UserController extends Controller
 {
@@ -15,6 +17,10 @@ class UserController extends Controller
     */
    public function index()
    {
+      if(!Auth::user()->hasPermissionTo('Super Usuario')) {
+         throw new UnauthorizedException('403', 'You do not have the required authorization!');
+      }
+
       $users = User::all();
 
       return view('root.users.index', [
@@ -29,6 +35,11 @@ class UserController extends Controller
     */
    public function create()
    {
+      if(!Auth::user()->hasPermissionTo('Super Usuario')) {
+         throw new UnauthorizedException('403', 'You do not have the required authorization!');
+      }
+
+
       return view('root.users.form', [
          'companies' => Company::all(['id', 'social_name']),
       ]);
@@ -42,6 +53,10 @@ class UserController extends Controller
     */
    public function store(Request $request)
    {
+      if(!Auth::user()->hasPermissionTo('Super Usuario')) {
+         throw new UnauthorizedException('403', 'You do not have the required authorization!');
+      }
+
       $user = User::create($request->all());
 
       return redirect()->route('root.users.edit', [
@@ -68,6 +83,10 @@ class UserController extends Controller
     */
    public function edit(User $user)
    {
+      if(!Auth::user()->hasPermissionTo('Super Usuario')) {
+         return back()->withToastWarning('Permissão negada!');
+      }
+
       return view('root.users.form', [
          'user' => $user,
          'companies' => Company::all(),
@@ -83,6 +102,10 @@ class UserController extends Controller
     */
    public function update(Request $request, User $user)
    {
+      if(!Auth::user()->hasPermissionTo('Super Usuario')) {
+         return back()->withToastWarning('Permissão negada!');
+      }
+
       $user->update($request->all());
 
       return redirect()->route('root.users.edit', [
@@ -98,6 +121,10 @@ class UserController extends Controller
     */
    public function destroy(User $user)
    {
+      if(!Auth::user()->hasPermissionTo('Super Usuario')) {
+         return back()->withToastWarning('Permissão negada!');
+      }
+
       $user->delete();
       return back()->withToastSuccess('Usuario removido com sucesso!');
    }
